Deprecate the theme prop of the ConfigurationProvider and render the theme root in the SlashIDProvider
