+++ packages/react/src/components/form/form.tsx
@@ -5,7 +5,6 @@ import { Initial } from "./initial";
 import { Authenticating } from "./authenticating";
 import { Error } from "./error";
 import { Success } from "./success";
-import { themeClass, darkTheme, autoTheme } from "../../theme/theme.css";
 import * as styles from "./form.css";
 import { Footer } from "./footer";
 import { useConfiguration } from "../../hooks/use-configuration";
@@ -28,20 +27,11 @@ export const Form: React.FC<Props> = ({
   text,
 }) => {
   const flowState = useFlowState({ onSuccess });
-  const { theme, showBanner } = useConfiguration();
+  const { showBanner } = useConfiguration();
   const { lastHandle } = useLastHandle();
 
   return (
-    <div
-      className={clsx(
-        "sid-theme-root",
-        `sid-theme-root__${theme}`,
-        themeClass,
-        { [darkTheme]: theme === "dark", [autoTheme]: theme === "auto" },
-        styles.form,
-        className
-      )}
-    >
+    <div className={clsx(styles.form, className)}>
       <ConfigurationOverrides text={text} factors={factors}>
         {flowState.status === "initial" && (
           <FormProvider>
