+++ packages/react/src/dev.tsx
@@ -70,7 +70,10 @@ function Config() {
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <SlashIDProvider oid={import.meta.env.VITE_ORG_ID}>
+    <SlashIDProvider
+      oid={import.meta.env.VITE_ORG_ID}
+      themeProps={{ theme: "light", className: "testClass" }}
+    >
       <Config />
     </SlashIDProvider>
   </React.StrictMode>
